Move localdb to api package
