+++ app/src/androidTest/java/com/rebalance/local/dao/ExpenseDaoTest.kt
@@ -4,13 +4,14 @@ import android.content.Context
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.rebalance.local.db.AppDatabase
-import com.rebalance.local.entities.Category
-import com.rebalance.local.entities.Expense
-import com.rebalance.local.entities.ExpenseUser
-import com.rebalance.local.entities.Group
-import com.rebalance.local.entities.User
-import com.rebalance.local.entities.UserGroup
+import com.rebalance.backend.localdb.dao.ExpenseDao
+import com.rebalance.backend.localdb.db.AppDatabase
+import com.rebalance.backend.localdb.entities.Category
+import com.rebalance.backend.localdb.entities.Expense
+import com.rebalance.backend.localdb.entities.ExpenseUser
+import com.rebalance.backend.localdb.entities.Group
+import com.rebalance.backend.localdb.entities.User
+import com.rebalance.backend.localdb.entities.UserGroup
 import junit.framework.TestCase.assertEquals
 import junit.framework.TestCase.assertNotNull
 import kotlinx.coroutines.runBlocking
