+++ app/src/main/java/com/rebalance/backend/localdb/entities/Category.kt
@@ -1,16 +1,13 @@
-package com.rebalance.local.entities
+package com.rebalance.backend.localdb.entities
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class User(
+data class Category(
     @PrimaryKey(autoGenerate = true) val id: Long?,
-    @ColumnInfo(name = "email") val email: String,
-    @ColumnInfo(name = "username") val username: String,
+    @ColumnInfo(name = "name") val name: String,
     @ColumnInfo(name = "db_id") val dbId: Long?,
     @ColumnInfo(name = "changed") val changed: Boolean
 )
-
-
