+++ app/src/main/java/com/rebalance/backend/localdb/entities/Group.kt
@@ -1,12 +1,13 @@
-package com.rebalance.local.entities
+package com.rebalance.backend.localdb.entities
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class Category(
+data class Group(
     @PrimaryKey(autoGenerate = true) val id: Long?,
+    @ColumnInfo(name = "currency") val currency: String,
     @ColumnInfo(name = "name") val name: String,
     @ColumnInfo(name = "db_id") val dbId: Long?,
     @ColumnInfo(name = "changed") val changed: Boolean
